+++ src/components/ResourceFormSuccess/ResourceFormSuccess.mobile.tsx
